+++ js/main.js
@@ -1,99 +1,7 @@
-const pictureCount = 25;
-const avatarCount = 6;
-const likeMinCount = 15;
-const likeMaxCount = 200;
-const commentCount = 20;
-const commentLines = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
+import {createMessage} from './data';
+import {createComment} from './data';
+import {createPictures} from './data';
 
-const descriptions = [
-  'Некоторые дни начинаются лучше остальных.',
-  'Сегодня в моей душе солнце.',
-  'Если ты ищешь того человека, который изменит твою жизнь, просто возьми и посмотри в зеркало.',
-  'Красота лишь привлекает внимание, душа – завоевывает сердце.',
-  'Все, что мы отдаем, возвращается вновь.',
-  'Цени моменты.',
-  'Улыбайся больше, смейся чаще.',
-  'Счастье никогда не выйдет из моды.',
-  'Они говорили мне, что я не смогу, поэтому я это сделала.',
-];
-
-const names = ['Роман', 'Яна', 'Джозеф', 'Карина', 'Нона', 'Тимур'];
-
-//функция случайного числа
-
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-//подфугкция - случайный элемент массива
-
-const getRandomArrayElement = (array) =>
-  array[getRandomInteger(0, array.length - 1)];
-
-
-//генератор для ID
-
-
-const createIdGenerator = () => {
-  let lastGeneratedId = 0;
-
-  return () => {
-    lastGeneratedId += 1;
-    return lastGeneratedId;
-  };
-};
-
-const generateCommentId = createIdGenerator();
-
-
-//функция создания сообщений
-
-const createMessage = () =>
-  Array.from({ length: getRandomInteger(1, 2) }, () =>
-    getRandomArrayElement(commentLines)
-  ).join(' ');
-
-
-//функция для создания одного комментария
-
-const createComment = () => ({
-  id: generateCommentId(),
-  avatar: `img/avatar-${getRandomInteger(1, avatarCount)}.svg`,
-  message: createMessage(),
-  name: getRandomArrayElement(names),
-});
-
-
-//функция, которая возвращает обьект
-
-const createPicture = (index) => ({
-  id: index,
-  url: `photos/${index}.jpg`,
-  descriptions: getRandomArrayElement(descriptions),
-  likes: getRandomInteger(likeMinCount, likeMaxCount),
-  comments: Array.from(
-    { length: getRandomInteger(0, commentCount) },
-    createComment
-  ),
-});
-
-
-//сначала генерируем 25 элементов массива, и далее функция, которая применяется к каждому элементу этого массива
-
-const createPictures = () =>
-  Array
-    .from({ length: pictureCount })
-    .map((_, pictureIndex) =>
-      createPicture(pictureIndex + 1));
-
+createMessage();
+createComment();
 createPictures();
